feat: modify profile and restriction that allows access for author only

+++ controllers/videoController.js
@@ -69,13 +69,17 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
+  req.user.videos.push(newVideo._id)
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
 // Video Detail
 
+// populate(객체를 데려오는 함수는 ObjectId에만 쓰일 수 있음.)
 export const videoDetail = async (req, res) => {
   const {
     params: {
@@ -83,7 +87,7 @@ export const videoDetail = async (req, res) => {
     }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate('creator');
     res.render("videoDetail", {
       pageTitle: video.title,
       video
@@ -103,10 +107,14 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", {
-      pageTitle: `Edit ${video.title}`,
-      video
-    });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", {
+        pageTitle: `Edit ${video.title}`,
+        video
+      });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -144,9 +152,13 @@ export const deleteVideo = async (req, res) => {
     }
   } = req;
   try {
-    await Video.findOneAndRemove({
-      _id: id
-    });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({
+        _id: id
+      });
+    }
   } catch (error) {
     console.error(error);
   }
